Update to handler contract.
Changes last handler in io_handlers to be a default handler.  Handler
address must still be 0.  Changes handler init to return BOOL
determining whether or not handler is ready to read/write data.

Also adds ability for handlers to change rdwr_data.autocommit to FALSE,
which allows for a handler to put the FX2  into GPIF or slave fifo mode
w/ auto in/out.